% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToCompactQP.R
\name{convertToCompact}
\alias{convertToCompact}
\title{Make a Constraint Matrix Compact}
\usage{
convertToCompact(Amat)
}
\arguments{
\item{Amat}{A constraint matrix as defined in solve.QP.}
}
\value{
A list containing the two elements to be passed to solve.QP.compact,
each named accordingly.
}
\description{
Make a Constraint Matrix Compact
}
\seealso{
quadprog::solve.QP

quadprog::solve.QP.compact
}
