% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeConstraints.R
\name{normalizeConstraints}
\alias{normalizeConstraints}
\title{Normalize constraint matrix}
\usage{
normalizeConstraints(Amat, bvec)
}
\arguments{
\item{Amat}{constraint matrix as defined by solve.QP}

\item{bvec}{constraints as defined by solve.QP}
}
\value{
a list with two elements: Amat and bvec that contain the normalized constraints.
}
\description{
it is not uncommon for quadprog to fail when there are large differences in 2-norm
between the columns of the constraint matrix (Amat).  It is possible to alleviate this issue in
some cases by normalizing the constraints (and their boundaries, defined by bvec).
}
\seealso{
quadprog::solve.QP

quadprog::solve.QP.compact
}
