% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToCompactQP.R
\name{convertToCompact}
\alias{convertToCompact}
\title{'Sparsify' constraint matrix}
\usage{
convertToCompact(Amat)
}
\arguments{
\item{Amat}{a constraint matrix as defined in solve.QP}
}
\value{
a list with two elements: Amat and Aind as necessary to be passed to solve.QP.compact
}
\description{
'Sparsify' constraint matrix
}
\seealso{
quadprog::solve.QP

quadprog::solve.QP.compact
}
